+++ packages/layout/src/columns/index.ts
@@ -1,0 +1,5 @@
+export { Columns } from './columns';
+
+// types
+
+export type { ColumnsProps } from './columns';
