+++ packages/layout/src/container/container.stories.tsx
@@ -1,6 +1,6 @@
 import type { ComponentMeta, ComponentStory } from '@storybook/react';
 
-import { Placeholder } from '../../../docs/components/example-helpers';
+import { Placeholder } from '../../../../docs/components/example-helpers';
 import type { ContainerProps } from './container';
 import { Container } from './container';
