+++ packages/layout/src/container/index.ts
@@ -1,0 +1,5 @@
+export { Container } from './container';
+
+// types
+
+export type { ContainerProps } from './container';
