+++ packages/layout/src/inline/index.ts
@@ -1,0 +1,5 @@
+export { Inline } from './inline';
+
+// types
+
+export type { InlineProps } from './inline';
