+++ packages/layout/src/stack/index.ts
@@ -1,0 +1,5 @@
+export { Stack } from './stack';
+
+// types
+
+export type { StackProps } from './stack';
