+++ packages/layout/src/stack/stack-alignment.ts
